LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/basic-shape-circle-ellipse-serialization-expected.txt


FAIL Serialization of basic shapes assert_equals: circle(at left bottom) expected "circle(at 0% 100%)" but got "circle(at left bottom)"
FAIL Serialization of basic shapes 1 assert_equals: circle(at bottom left) expected "circle(at 0% 100%)" but got "circle(at left bottom)"
FAIL Serialization of basic shapes 2 assert_equals: circle(at right calc(10% + 5px)) expected "circle(at 100% calc(10% + 5px))" but got "circle(at right calc(10% + 5px))"
FAIL Serialization of basic shapes 3 assert_equals: ellipse(at left bottom) expected "ellipse(at 0% 100%)" but got "ellipse(at left bottom)"
FAIL Serialization of basic shapes 4 assert_equals: ellipse(at bottom left) expected "ellipse(at 0% 100%)" but got "ellipse(at left bottom)"
FAIL Serialization of basic shapes 5 assert_equals: ellipse(at right calc(10% + 5px)) expected "ellipse(at 100% calc(10% + 5px))" but got "ellipse(at right calc(10% + 5px))"
PASS Serialization of basic shapes 6
PASS Serialization of basic shapes 7
PASS Serialization of basic shapes 8
PASS Serialization of basic shapes 9
PASS Serialization of basic shapes 10
PASS Serialization of basic shapes 11
FAIL Serialization of basic shapes 12 assert_equals: circle(at top 0% right 5px) expected "circle(at calc(100% - 5px) 0%)" but got "circle(at right 5px top 0%)"
FAIL Serialization of basic shapes 13 assert_equals: ellipse(at top 0% right 10px) expected "ellipse(at calc(100% - 10px) 0%)" but got "ellipse(at right 10px top 0%)"
FAIL Serialization of basic shapes 14 assert_equals: circle(closest-side at center) expected "circle(at 50% 50%)" but got "circle(at center center)"
FAIL Serialization of basic shapes 15 assert_equals: ellipse(closest-side closest-side at center) expected "ellipse(at 50% 50%)" but got "ellipse(at center center)"
FAIL Serialization of basic shapes 16 assert_equals: circle(farthest-side at center) expected "circle(farthest-side at 50% 50%)" but got "circle(farthest-side at center center)"
FAIL Serialization of basic shapes 17 assert_equals: circle(10px at center) expected "circle(10px at 50% 50%)" but got "circle(10px at center center)"
FAIL Serialization of basic shapes 18 assert_equals: ellipse(farthest-side 10px at center) expected "ellipse(farthest-side 10px at 50% 50%)" but got "ellipse(farthest-side 10px at center center)"
PASS Serialization of basic shapes 19
PASS Serialization of basic shapes 20
FAIL Serialization of basic shapes 21 assert_equals: circle(at right 5px bottom 10px) expected "circle(at calc(100% - 5px) calc(100% - 10px))" but got "circle(at right 5px bottom 10px)"
FAIL Serialization of basic shapes 22 assert_equals: ellipse(at right 5px bottom 10px) expected "ellipse(at calc(100% - 5px) calc(100% - 10px))" but got "ellipse(at right 5px bottom 10px)"
FAIL Serialization of basic shapes 23 assert_equals: circle(at right 5% top 0px) expected "circle(at calc(95%) 0px)" but got "circle(at right 5% top 0px)"
FAIL Serialization of basic shapes 24 assert_equals: ellipse(at right 5% top 0px) expected "ellipse(at calc(95%) 0px)" but got "ellipse(at right 5% top 0px)"
FAIL Serialization of basic shapes 25 assert_equals: circle(at right calc(10% + 5px) bottom calc(10% + 5px)) expected "circle(at calc(90% - 5px) calc(90% - 5px))" but got "circle(at right calc(10% + 5px) bottom calc(10% + 5px))"
FAIL Serialization of basic shapes 26 assert_equals: ellipse(at right calc(10% + 5px) bottom calc(10% + 5px)) expected "ellipse(at calc(90% - 5px) calc(90% - 5px))" but got "ellipse(at right calc(10% + 5px) bottom calc(10% + 5px))"
